requirements.txt
albumentations==1.3.0
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.0.4
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
fire==0.4.0
fonttools==4.37.4
idna==3.4
imageio==2.22.1
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.1
nbimporter==0.3.4
networkx==2.8.7
numpy==1.23.4
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
pybboxes==0.1.5
pycodestyle==2.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
sahi==0.10.7
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.2
Shapely==1.8.5.post1
six==1.16.0
termcolor==2.0.1
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2022.10.10
toml==0.10.2
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
